Fixed DRY in RecyclerViewUis
