+++ app/src/androidTest/java/com/veselovvv/coinsapp/rates/RatesRecyclerViewUi.kt
@@ -5,11 +5,11 @@ import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.contrib.RecyclerViewActions.scrollToPosition
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import com.veselovvv.coinsapp.R
-import com.veselovvv.coinsapp.core.RecyclerViewUi
+import com.veselovvv.coinsapp.core.NoResultsRecyclerViewUi
 import com.veselovvv.coinsapp.core.withItemViewType
 import com.veselovvv.coinsapp.core.withRecyclerViewItemText
 
-class RatesRecyclerViewUi : RecyclerViewUi(
+class RatesRecyclerViewUi : NoResultsRecyclerViewUi(
     R.id.rates_swipe_to_refresh,
     R.id.rates_recycler_view
 ) {
@@ -24,19 +24,10 @@ class RatesRecyclerViewUi : RecyclerViewUi(
         }
     }
 
-    //TODO DRY
-    fun checkNoResultsState(text: String) {
-        interaction.perform(scrollToPosition<RecyclerView.ViewHolder>(0))
-            .check(matches(withItemViewType(NO_RESULTS_VIEW_TYPE)))
-            .check(matches(isDisplayed()))
-            .check(matches(withRecyclerViewItemText(R.id.no_results_text_view, text)))
-    }
-
     /**
      * Described in getItemViewType() in RatesAdapter
      */
     companion object {
-        private const val NO_RESULTS_VIEW_TYPE = -1
         private const val BASE_VIEW_TYPE = 0
     }
 }
